Refactor: Chat Layout 폴더 구조 개선
